requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-dotenv==0.19.0
requests==2.26.0
pandas==1.3.3
openpyxl==3.0.9
streamlit==1.2.0
plotly==5.3.1
schedule==1.1.0 